+++ public/index.html
@@ -18,12 +18,10 @@
           <div class="trip-main__trip-controls  trip-controls">
             <div class="trip-controls__navigation">
               <h2 class="visually-hidden">Switch trip view</h2>
-              <!-- Меню -->
             </div>
 
             <div class="trip-controls__filters">
               <h2 class="visually-hidden">Filter events</h2>
-              <!-- Фильтры -->
             </div>
           </div>
 
@@ -37,13 +35,7 @@
             <h2 class="visually-hidden">Trip events</h2>
             <ul class="trip-events__list">
 
-
-
             </ul>
-
-            <!-- Сортировка -->
-
-            <!-- Контент -->
           </section>
       </div>
     </main>
