+++ src/main.js
@@ -1,19 +1,12 @@
-import {listSorting, sortingEvents} from './view/list-sorting.js';
-import { filteringList } from './view/list-filter.js';
-import { pointList } from './view/list-content.js';
-import { tripNavigation } from './view/menu.js';
-import { editTrip } from './view/edit-trip.js';
-import { tripInfo } from './view/trip-info.js';
+import { renderTemplate, renderPos } from 'render.js';
+import { createListSort} from './view/list-sorting.js';
+import { createListFilter } from './view/list-filter.js';
+import { createPointList } from './view/list-content.js';
+import { createTripNav } from './view/menu.js';
+import { createEditTrip } from './view/edit-trip.js';
+import { createTripInfo } from './view/trip-info.js';
 
-export const renderPos = {
-  BEFOREBEGIN: 'beforebegin',
-  AFTERBEGIN: 'afterbegin',
-  BEFOREEND: 'beforeend',
-  AFTEREND: 'afterend',
-};
-const renderTemplate = (container, template, placement) => {
-  container.insertAdjacentHTML(placement, template);
-};
+const TRIP_POINT_NUM = 3;
 
 
 const tripMenu = document.querySelector('.trip-controls__navigation');
@@ -24,14 +17,14 @@ const tripEvents = document.querySelector('.trip-events');
 const tripPointsList = document.querySelector('.trip-events__list');
 
 
-renderTemplate(tripMain, tripInfo(), renderPos.AFTERBEGIN);
-renderTemplate(tripMenu, tripNavigation(), renderPos.BEFOREEND);
-renderTemplate(tripControlsFilter, filteringList(), renderPos.BEFOREEND);
+renderTemplate(tripMain, createTripInfo(), renderPos.AFTERBEGIN);
+renderTemplate(tripMenu, createTripNav(), renderPos.BEFOREEND);
+renderTemplate(tripControlsFilter, createListFilter(), renderPos.BEFOREEND);
 
-renderTemplate(tripEvents, listSorting(), renderPos.AFTERBEGIN);
-renderTemplate(tripPointsList, editTrip(), renderPos.AFTERBEGIN);
-for(let i = 0; i < 3; i++){
-  renderTemplate(tripPointsList, pointList(), renderPos.BEFOREEND);
+renderTemplate(tripEvents, createListSort(), renderPos.AFTERBEGIN);
+renderTemplate(tripPointsList, createEditTrip(), renderPos.AFTERBEGIN);
+for(let i = 0; i < TRIP_POINT_NUM; i++){
+  renderTemplate(tripPointsList, createPointList(), renderPos.BEFOREEND);
 }
