+++ src/view/edit-trip.js
@@ -1,4 +1,4 @@
-export const editTrip = () =>
+export const createEditTrip = () =>
   (`<li class="trip-events__item">
 <form class="event event--edit" action="#" method="post">
   <header class="event__header">
