+++ src/view/list-content.js
@@ -1,4 +1,4 @@
-export const pointList = () => (
+export const createPointList = () => (
   `       <li class="trip-events__item">
               <div class="event">
                 <time class="event__date" datetime="2019-03-18">MAR 18</time>
