+++ src/view/list-filter.js
@@ -1,4 +1,4 @@
-export const filteringList = () => (
+export const createListFilter = () => (
   `<form class="trip-filters" action="#" method="get">
 	<div class="trip-filters__filter">
 	  <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything">
