+++ src/view/menu.js
@@ -1,4 +1,4 @@
-export const tripNavigation = () =>
+export const createTripNav = () =>
   (`
 	<nav class="trip-controls__trip-tabs  trip-tabs">
 		<a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
