+++ src/view/trip-info.js
@@ -1,4 +1,4 @@
-export const tripInfo = () =>
+export const createTripInfo = () =>
   (`
   <section class="trip-main__trip-info  trip-info">
   <div class="trip-info__main">
